api_yamdb/requirements.txt
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
django-environ==0.10.0
gunicorn==20.0.4
psycopg2-binary==2.8.6